+++ CHANGES.md
@@ -1,6 +1,25 @@
 twilio-java changelog
 =====================
 
+[2022-03-09] Version 8.27.1
+---------------------------
+**Library - Chore**
+- [PR #672](https://github.com/twilio/twilio-java/pull/672): push Datadog Release Metric upon deploy success. Thanks to [@eshanholtz](https://github.com/eshanholtz)!
+
+**Api**
+- Add optional boolean include_soft_deleted parameter to retrieve soft deleted recordings
+
+**Chat**
+- Add `X-Twilio-Wehook-Enabled` header to `delete` method in UserChannel resource
+
+**Numbers**
+- Expose `failure_reason` in the Supporting Documents resources
+
+**Verify**
+- Add optional `metadata` parameter to "verify challenge" endpoint, so the SDK/App can attach relevant information from the device when responding to challenges.
+- remove beta feature flag to list atempt api operations.
+
+
 [2022-02-23] Version 8.27.0
 ---------------------------
 **Library - Chore**